refactor code

+++ ship-oa/src/main/java/com/yodean/oa/common/core/service/BaseService.java
@@ -1,11 +1,15 @@
 package com.yodean.oa.common.core.service;
 
 import com.google.common.collect.Lists;
+import com.rick.db.service.JdbcService;
 import com.yodean.common.enums.DelFlag;
 import com.yodean.oa.common.exception.OAException;
 import com.yodean.oa.common.plugin.document.enums.ExceptionCode;
 import com.yodean.platform.api.util.EntityBeanUtils;
 import com.yodean.platform.domain.BaseEntity;
+import org.apache.commons.lang3.ArrayUtils;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.*;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import javax.transaction.Transactional;
@@ -17,26 +21,26 @@ import java.util.Optional;
 /**
  * Created by rick on 7/19/18.
  */
-public abstract class BaseService<T extends BaseEntity> {
+public abstract class BaseService<T extends BaseEntity, R extends JpaRepository<T, Long>> {
 
-    abstract protected JpaRepository<T, Long> autowired();
+    @Autowired
+    protected R jpaRepository;
+
+    @Autowired
+    protected JdbcService jdbcService;
 
     /**
-     * 增量保存，忽略null的字段
+     * 添加或编辑
      * @param t
      * @return
      */
     @Transactional
     public T save(T t) {
-        JpaRepository<T,Long> jpaRepository = autowired();
         return jpaRepository.save(merge(t));
     }
 
     @Transactional
     public List<T> saveAll(Collection<T> list) {
-
-        JpaRepository<T,Long> jpaRepository = autowired();
-
         List<T> saveList = Lists.newArrayListWithExpectedSize(list.size());
 
         list.forEach(t -> saveList.add(merge(t)));
@@ -48,7 +52,7 @@ public abstract class BaseService<T extends BaseEntity> {
         //参数格式化
         t.initParams();
 
-        if (Objects.nonNull(t.getId())) { //修改
+        if (Objects.nonNull(t.getId())) { //修改,忽略null字段
             T persist = findById(t.getId());
 
             EntityBeanUtils.merge(persist, t);
@@ -65,8 +69,6 @@ public abstract class BaseService<T extends BaseEntity> {
      * @return
      */
     public T findById(Long id) {
-        JpaRepository<T,Long> jpaRepository = autowired();
-
         Optional<T> optional = jpaRepository.findById(id);
         if (!optional.isPresent()) {
             throw new OAException(ExceptionCode.NOT_FOUND_ERROR);
@@ -89,39 +91,67 @@ public abstract class BaseService<T extends BaseEntity> {
      * @return
      */
     public T getOne(Long id) {
-        JpaRepository<T,Long> jpaRepository = autowired();
         return jpaRepository.getOne(id);
     }
 
     /**
-     * 根据id删除
-     * @param id
+     * 根据ids物理删除
+     * @param ids
      * @return
      */
     @Transactional
-    public void delete(Long id) {
-        JpaRepository<T, Long> jpaRepository = autowired();
-        jpaRepository.deleteById(id);
-    }
+    public void delete(Long... ids) {
+        if (ArrayUtils.isEmpty(ids))
+            return;
 
-    public void deleteAll(Collection<T> list) {
-        JpaRepository<T, Long> jpaRepository = autowired();
-        jpaRepository.deleteInBatch(list);
+        if (ids.length == 1) {
+            jpaRepository.deleteById(ids[0]);
+        } else { //批量删除
+            List<T> list = Lists.newArrayListWithCapacity(ids.length);
+
+            for (Long id : ids) {
+                jpaRepository.findById(id).ifPresent(list::add);
+            }
+            jpaRepository.deleteInBatch(list);
+        }
     }
 
     /**
      * 根据id逻辑删除
-     * @param id
+     * @param ids
      * @return
      */
     @Transactional
-    public void deleteByFlag(Long id) {
-        JpaRepository<T,Long> jpaRepository = autowired();
-        Optional<T> optional = jpaRepository.findById(id);
-        if (optional.isPresent()) {
-            T t = optional.get();
-            t.setDelFlag(DelFlag.DEL_FLAG_REMOVE);
-            jpaRepository.save(t);
+    public void deleteByFlag(Long... ids) {
+        if (ArrayUtils.isEmpty(ids))
+            return;
+
+        for (Long id : ids) {
+            Optional<T> optional = jpaRepository.findById(id);
+            if (optional.isPresent()) {
+                T t = optional.get();
+                t.setDelFlag(DelFlag.DEL_FLAG_REMOVE);
+                jpaRepository.save(t);
+            }
         }
+    }
+
+    /**
+     * 分页(只做Example的查询分页)
+     * 如果做JpaSpecificationExecutor的分页需要自己写特定的方法
+     * 根据业务自己实现
+     * Note:起始页一定要注意是从0开始的，而不是从1开始
+     */
+    public Page<T> findAll(T t, int page, int size, Sort.Direction direction, String... properties ) {
+
+
+//        ExampleMatcher matcher = ExampleMatcher.matching()
+//         .withMatcher("title", ExampleMatcher.GenericPropertyMatchers.contains()) ;
+//
+//        Example<T> example = Example.of(t, matcher);
+
+        Example<T> example = Example.of(t);
+        Pageable pageable = PageRequest.of(page, size, direction, properties);
+        return jpaRepository.findAll(example, pageable);
     }
 }
